# Project-Mannager-System/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PM Daily Work Management API
  version: 1.0.0
  description: |
    API cho ứng dụng quản lý công việc hằng ngày của quản lý dự án phần mềm.

    ## Quy ước chung
    - Base path: `/api/v1`
    - Xác thực: `Authorization: Bearer <accessToken>` (JWT HS256, hết hạn 15 phút). Refresh token (7 ngày) chỉ gửi ở `/auth/refresh` và `/auth/logout`.
    - Thời gian: ISO-8601 UTC (`timestamptz`). Ngày (date): `yyyy-MM-dd`.
    - Phân trang: `page` (≥ 0, mặc định 0), `size` (1–100, mặc định 20), `sort` (whitelist theo từng endpoint, VD `fullName,asc`).
    - Lỗi: format thống nhất (BR-GEN-05) — `timestamp, status, error, code, message, path, fieldErrors, traceId`. Catalog mã lỗi: `docs/design/05-error-handling-design.md`.
    - Dữ liệu đã xóa mềm: không xuất hiện ở danh sách; truy cập trực tiếp trả 404.
    - Cập nhật bắt buộc kèm `version`; version cũ → 409.
    - Chi tiết từng module: `docs/api/01..12-*.md`.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: 10 nhóm số liệu tổng quan hằng ngày
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: fromDate, in: query, schema: { type: string, format: date } }
        - { name: toDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Tổng quan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardSummaryResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /dashboard/task-stats:
    get:
      tags: [Dashboard]
      summary: Biểu đồ tasksByStatus + tasksByPriority
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: fromDate, in: query, schema: { type: string, format: date } }
        - { name: toDate, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Thống kê
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardTaskStatsResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /dashboard/projects/progress:
    get:
      tags: [Dashboard]
      summary: Biểu đồ tiến độ dự án
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQuery' }
      responses:
        '200':
          description: Tiến độ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardProjectsProgressResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  # ============================= PROJECT =============================
...
  /action-items:
    get:
      tags: [Action Item]
      summary: Danh sách action item
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ActionItemStatus' } }
        - { name: assigneeId, in: query, schema: { type: string, format: uuid } }
        - { name: overdue, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItemPage' }
        '403': { $ref: '#/components/responses/Forbidden403' }
    post:
      tags: [Action Item]
      summary: Tạo action item từ cuộc họp (sinh ACTION_ITEM_ASSIGNED)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionItemCreateRequest' }
      responses:
        '201':
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItemResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /action-items/overdue:
    get:
      tags: [Action Item]
      summary: Action item quá hạn chưa đóng
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItemPage' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /action-items/{id}:
    get:
      tags: [Action Item]
      summary: Chi tiết action item
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ActionItemIdPath' }
      responses:
        '200':
          description: Chi tiết
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItemResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    put:
      tags: [Action Item]
      summary: "Cập nhật (assignee: chỉ status/progress)"
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ActionItemIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionItemUpdateRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItemResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409': { $ref: '#/components/responses/Conflict409' }
    delete:
      tags: [Action Item]
      summary: Xóa action item
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ActionItemIdPath' }
      responses:
        '204': { description: Thành công, không body }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /action-items/{id}/convert-to-task:
    post:
      tags: [Action Item]
      summary: Chuyển action item thành task (liên kết 1–1)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ActionItemIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConvertToTaskRequest' }
      responses:
        '201':
          description: Task được tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409':
          description: ALREADY_LINKED — action item đã có linked task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  # ============================= RISK =============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Số trang (từ 0)
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Số bản ghi mỗi trang
    Sort:
      name: sort
      in: query
      schema: { type: string }
      description: 'Sắp xếp theo whitelist của endpoint, VD `fullName,asc`'
...
    ProjectIdQuery:
      name: projectId
      in: query
      schema: { type: string, format: uuid }
      description: Lọc theo dự án
...
    ActionItemIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: ID action item
...
  responses:
    Validation400:
      description: Dữ liệu không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden403:
      description: Không đủ quyền hoặc ngoài phạm vi dữ liệu
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound404:
      description: Không tìm thấy (bao gồm dữ liệu đã xóa mềm)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict409:
      description: Xung đột optimistic locking (version cũ)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path, fieldErrors, traceId]
      properties:
        timestamp: { type: string, format: date-time, example: '2026-08-01T10:00:00Z' }
        status: { type: integer, example: 400 }
        error: { type: string, example: BAD_REQUEST }
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: 'Dữ liệu không hợp lệ' }
        path: { type: string, example: /api/v1/tasks }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        traceId: { type: string, example: a1b2c3d4-1234 }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
    # ---------- Enum ----------
...
    TaskStatus: { type: string, enum: [TODO, IN_PROGRESS, BLOCKED, REVIEW, DONE, CANCELLED] }
    TaskPriority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    TaskType: { type: string, enum: [FEATURE, BUG, IMPROVEMENT, TASK, OTHER] }
    TaskSource: { type: string, enum: [MANUAL, MEETING, ACTION_ITEM, ISSUE, OTHER] }
...
    ActionItemStatus: { type: string, enum: [OPEN, IN_PROGRESS, DONE, CANCELLED] }
...
    DashboardSummaryResponse:
      type: object
      properties:
        totalTasksToday: { type: integer }
        overdueTasks: { type: integer }
        upcomingTasks: { type: integer }
        inProgressTasks: { type: integer }
        blockedTasks: { type: integer }
        meetingsToday: { type: integer }
        pendingActionItems: { type: integer }
        highRisks: { type: integer }
        openIssues: { type: integer }
        upcomingMilestones: { type: integer }
    DashboardTaskStatsResponse:
      type: object
      properties:
        tasksByStatus:
          type: array
          items:
            type: object
            properties:
              status: { $ref: '#/components/schemas/TaskStatus' }
              count: { type: integer }
        tasksByPriority:
          type: array
          items:
            type: object
            properties:
              priority: { $ref: '#/components/schemas/TaskPriority' }
              count: { type: integer }
    DashboardProjectsProgressResponse:
      type: object
      properties:
        projects:
          type: array
          items: { $ref: '#/components/schemas/ProjectProgressItem' }
    ProjectProgressItem:
      type: object
      properties:
        projectId: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
    # ---------- Project ----------
...
    TaskResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PRJ001-TASK-000001 }
        projectId: { type: string, format: uuid }
        projectCode: { type: string }
        projectName: { type: string }
        parentTaskId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string }
        reporter: { $ref: '#/components/schemas/UserRef' }
        assignee: { $ref: '#/components/schemas/UserRef' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        type: { $ref: '#/components/schemas/TaskType' }
        source: { $ref: '#/components/schemas/TaskSource' }
        startDate: { type: string, format: date, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        actualCompletedAt: { type: string, format: date-time, nullable: true }
        progress: { type: integer }
        blocked: { type: boolean }
        blockerReason: { type: string, nullable: true }
        estimateMinutes: { type: integer, nullable: true }
        actualMinutes: { type: integer, nullable: true }
        notes: { type: string }
        tags:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              color: { type: string }
        collaborators:
          type: array
          items: { $ref: '#/components/schemas/UserRef' }
        watchers:
          type: array
          items: { $ref: '#/components/schemas/UserRef' }
        commentCount: { type: integer }
        attachmentCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
...
    ActionItemCreateRequest:
      type: object
      required: [meetingId, projectId, title, assigneeId]
      properties:
        meetingId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string }
        assigneeId: { type: string, format: uuid }
        dueDate: { type: string, format: date }
        priority: { $ref: '#/components/schemas/TaskPriority' }
    ActionItemUpdateRequest:
      type: object
      required: [version]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string }
        dueDate: { type: string, format: date }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        status: { $ref: '#/components/schemas/ActionItemStatus' }
        progress: { type: integer, minimum: 0, maximum: 100 }
        version: { type: integer }
    ConvertToTaskRequest:
      type: object
      properties:
        dueDate: { type: string, format: date }
        priority: { $ref: '#/components/schemas/TaskPriority' }
    ActionItemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        meetingId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        assignee: { $ref: '#/components/schemas/UserRef' }
        dueDate: { type: string, format: date, nullable: true }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        status: { $ref: '#/components/schemas/ActionItemStatus' }
        progress: { type: integer }
        linkedTaskId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        version: { type: integer }
...
    ActionItemPage:
      allOf:
        - { $ref: '#/components/schemas/PageResponse' }
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ActionItemResponse' }
    # ---------- Risk ----------
...